feat(security): ENG-2227 add key-gen module and add disk encryption for AKS (#80)
Adds the ability to enable disk encryption to an AKS cluster by
providing a `disk_encryption_set_id` to the AKS module. This is provided
by the new `key-gen` module which creates a key in a key vault. It will
aditionally create an encryption_set for the given key when the value is
enabled. This module was previously integrated into the `key-vault`
module but has now been separated out.